refactor: move gateway onboarding into extensions
